remove api file in the Profile page and inline the callbacks
